Clean up main.js by extracting functions into modules
Moved game board, mode switching, and timer functions to separate files to make main.js more organized and easier to maintain.
